requirements.txt
# FlowBrain — Python dependencies
# Install with: pip install -r requirements.txt

# Core: vector database (local, no server needed)
chromadb>=0.5.0

# Sentence embeddings — runs locally, no API key needed
sentence-transformers>=3.0.0

# Web server
fastapi>=0.111.0
uvicorn[standard]>=0.30.0

# HTTP client (for calling n8n webhooks)
httpx>=0.27.0

# Data handling
pydantic>=2.0.0

# Progress bars (for indexer)
tqdm>=4.66.0

# HTTP requests (for harvester)
requests>=2.32.0

# Environment variable loading
python-dotenv>=1.0.0

# ── Test / dev dependencies ──────────────────────────────────────────────────
pytest>=8.0.0
